render scmm project settings tab
