+++ app/views/projects/settings/_scmm.html.erb
@@ -1,0 +1,1 @@
+Hello World
