+++ config/locales/en.yml
@@ -1,3 +1,3 @@
 # English strings go here for Rails i18n
 en:
-  # my_label: "My label"
+  scmm: "SCM-Manager"
